Rename "general-schema-validator.ts" to "validate.ts"
