+++ src/index.ts
@@ -1,3 +1,5 @@
+export { ERROR_CODE } from './error'
+
 export {
   string,
   number,
@@ -9,13 +11,10 @@ export {
 } from './programmatic-schema'
 
 export { error, data } from './utils/fp'
-
 export { parse } from './general-schema-parser'
-export { validate, guard } from './general-schema-validator'
-export { ERROR_CODE } from './error'
+export { validate, guard } from './validate'
 
-export type { BaseSchema, Schema, SubjectType } from './types/compounds'
+export type { Schema, SubjectType } from './types/compounds'
 
 export type { InvalidSubject, ErrorPath, ErrorCode } from './error'
-
 export type { EitherError, Error, Data } from './utils/fp'
