+++ src/programmatic-schema.ts
@@ -1,4 +1,4 @@
-import { validate } from './general-schema-validator'
+import { validate } from './validate'
 import { parse } from './general-schema-parser'
 
 import type { EitherError } from './utils/fp'
