+++ src/test/programmatic-schema/number.test.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../general-schema-parser'
-import { validate } from '../../general-schema-validator'
+import { validate } from '../../validate'
 import { number } from '../../programmatic-schema'
 import { check, unknownX } from '../test-utils'
