+++ src/test/programmatic-schema/string.test.ts
@@ -1,6 +1,6 @@
 import { ERROR_CODE } from '../../error'
 import { parse } from '../../general-schema-parser'
-import { validate } from '../../general-schema-validator'
+import { validate } from '../../validate'
 import { string, makeStruct } from '../../programmatic-schema'
 import { check } from '../test-utils'
