+++ src/validate.ts
